feat(cpu): add RadixAttnOp and Scatter2ShardsOp support
- Register `RadixAttnOp` and `Scatter2ShardsOp` in CPU backend
- Implement forward logic for `RadixAttnOp` with architecture-specific kernels
- Update `RadixAttnOpOptions` to include head count parameters
- Fix const-correctness and template usage in radix attention kernel
- Remove unnecessary `.to(kFloat16)` calls in Qwen3 attention module
- Adjust function signatures for better type safety in radix attention kernel
- Refactor kernel calls to use namespaced static dispatch

+++ mllm/models/qwen3/modeling_qwen3_service.hpp
@@ -174,11 +174,11 @@ class Qwen3Attention final : public nn::Module {
     // [B, H, S, D]
     query_states = query_states.transpose(1, 2);
     key_states = key_states.transpose(1, 2);
-    value_states = value_states.transpose(1, 2).to(kFloat16);
+    value_states = value_states.transpose(1, 2);
 
     // [B, H, S, D]
-    query_states = q_rope_(query_states, llm_embedding_sin, llm_embedding_cos).to(kFloat16);
-    key_states = k_rope_(key_states, llm_embedding_sin, llm_embedding_cos).to(kFloat16);
+    query_states = q_rope_(query_states, llm_embedding_sin, llm_embedding_cos);
+    key_states = k_rope_(key_states, llm_embedding_sin, llm_embedding_cos);
 
     // FIXME: Think, if rope before cache is ok?
